Create ATM INTERFACE
This complex project consists of
five different classes and is a console-based application. When the system starts the user is
prompted with user id and user pin. On entering the details successfully, then ATM functionalities
are unlocked. The project allows to perform following operations:

Transactions History
Withdraw
Deposit
Transfer
Quit